Make apiclient a package
Splits apiclient.py into apiclient/ errors, base, data and search.
Keeps everything imported by other projects in dmutils.apiclient
namespace.
